Better UX around notebook deletion

+++ src/Sidebar.js
@@ -2,19 +2,31 @@ import React, { useContext, useEffect } from 'react';
 import styled from 'styled-components';
 
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-import { faPlus, faTrash, faPencilAlt } from '@fortawesome/free-solid-svg-icons'
+import {
+  faBook,
+  faPlus,
+  faTrash,
+  faPencilAlt
+} from '@fortawesome/free-solid-svg-icons';
+import { wrapComponent } from 'react-snackbar-alert';
 
 import Notebook from './Notebook';
 
 import * as ipc from './ipc';
 
-import { SET_CURRENT_NOTEBOOK, SET_NOTEBOOKS, SET_NOTES, ADD_NOTEBOOK, DELETE_NOTEBOOK } from './store/notes';
+import {
+  SET_CURRENT_NOTEBOOK,
+  SET_NOTEBOOKS,
+  SET_NOTES,
+  ADD_NOTEBOOK,
+  DELETE_NOTEBOOK
+} from './store/notes';
 import { SET_RENAMING_NOTEBOOK } from './store/app';
 import { App, Notes } from './store';
 
 const Container = styled.div`
-  background: ${({theme}) => theme.sidebarBackground};
-  color: #FFFFFF;
+  background: ${({ theme }) => theme.sidebarBackground};
+  color: #ffffff;
   height: 100%;
   user-select: none;
 
@@ -47,19 +59,28 @@ const Toolbar = styled.div`
     outline: none;
 
     &:hover {
-      color: #FFFFFF;
+      color: #ffffff;
+    }
+
+    &:disabled {
+      opacity: 0.5;
+      cursor: not-allowed;
+
+      &:hover {
+        color: rgba(255, 255, 255, 0.5);
+      }
     }
   }
 `;
 
-export default function Sidebar() {
+function Sidebar({ createSnackbar }) {
   const { notebooks, currentNotebook } = useContext(Notes.State);
   const notesDispatch = useContext(Notes.Dispatch);
 
   const appDispatch = useContext(App.Dispatch);
 
   function onClickNotebook(notebook) {
-    notesDispatch({ type: SET_CURRENT_NOTEBOOK, payload: notebook }); 
+    notesDispatch({ type: SET_CURRENT_NOTEBOOK, payload: notebook });
   }
 
   async function onClickNew() {
@@ -68,8 +89,19 @@ export default function Sidebar() {
   }
 
   async function onClickDelete() {
-    await ipc.deleteNotebook(currentNotebook.name);
-    notesDispatch({ type: DELETE_NOTEBOOK });
+    const confirm = await ipc.confirmDeleteNotebook(currentNotebook.name);
+    if (confirm) {
+      await ipc.deleteNotebook(currentNotebook.name);
+      notesDispatch({ type: DELETE_NOTEBOOK });
+      createSnackbar({
+        message: `Notebook "${currentNotebook.name}" was deleted.`,
+        theme: 'success'
+      });
+
+      if (notebooks.length === 1) { // we deleted the last notebook
+        notesDispatch({ type: SET_NOTES, payload: [] });
+      }
+    }
   }
 
   function onClickRename() {
@@ -103,22 +135,71 @@ export default function Sidebar() {
         <button onClick={onClickNew} data-tip="New Notebook">
           <FontAwesomeIcon icon={faPlus} />
         </button>
-        <button onClick={onClickRename} data-tip="Rename Notebook">
+        <button
+          onClick={onClickRename}
+          data-tip="Rename Notebook"
+          disabled={!notebooks.length}
+        >
           <FontAwesomeIcon icon={faPencilAlt} />
         </button>
-        <button onClick={onClickDelete} data-tip="Delete Notebook">
+        <button
+          onClick={onClickDelete}
+          data-tip="Delete Notebook"
+          disabled={!notebooks.length}
+        >
           <FontAwesomeIcon icon={faTrash} />
         </button>
-      </Toolbar>    
-      <NotebookList>
-        {notebooks.map(notebook => (
-          <Notebook
-            key={notebook.id}
-            onClick={onClickNotebook}
-            notebook={notebook}
-            active={currentNotebook.id === notebook.id} />
-        ))}
-      </NotebookList>
+      </Toolbar>
+      {notebooks && notebooks.length ? (
+        <NotebookList>
+          {notebooks.map(notebook => (
+            <Notebook
+              key={notebook.id}
+              onClick={onClickNotebook}
+              notebook={notebook}
+              active={currentNotebook.id === notebook.id}
+            />
+          ))}
+        </NotebookList>
+      ) : (
+        <NoNotebooksMessage />
+      )}
     </Container>
   );
 }
+
+const StyledMessage = styled.div`
+  text-align: center;
+  align-self: center;
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+  align-items: center;
+  justify-content: center;
+
+  h3 {
+    font-size: 1.1rem;
+    margin: 0;
+  }
+`;
+
+const NoNotebooksIcon = styled.div`
+  font-size: 1.5rem;
+  opacity: 0.2;
+`;
+
+function NoNotebooksMessage() {
+  return (
+    <StyledMessage>
+      <NoNotebooksIcon>
+        <FontAwesomeIcon icon={faBook} />
+      </NoNotebooksIcon>
+      <h3>There are no notebooks.</h3>
+      <p>
+        Click the <FontAwesomeIcon icon={faPlus} /> to get started.
+      </p>
+    </StyledMessage>
+  );
+}
+
+export default wrapComponent(Sidebar);
